Implementation of creature unit test

+++ data/creature.py
@@ -1,3 +1,6 @@
+from sqlite3 import IntegrityError
+
+from errors import Duplicate, Missing
 from .init import conn, curs
 from model.creature import Creature
 
@@ -19,7 +22,11 @@ def get_one(name: str) -> Creature:
     qry = "select * from creature where name=:name"
     params = {"name": name}
     curs.execute(qry, params)
-    return row_to_model(curs.fetchone())
+    row = curs.fetchone()
+    if row:
+        return row_to_model(row)
+    else:
+        raise Missing(msg=f"Explorer {name} not found")
 
 def get_all() -> list[Creature]:
     qry = "select * from creature"
@@ -30,26 +37,35 @@ def create(creature: Creature) -> Creature:
     qry = """insert into creature values
           (:name, :description, :country, :area, :aka)"""
     params = model_to_dict(creature)
-    curs.execute(qry, params)
-    conn.commit()
+    try:
+        curs.execute(qry, params)
+        conn.commit()
+    except IntegrityError:
+        raise Duplicate(msg=f"Explorer {creature.name} already exists")
     return get_one(creature.name)
 
-def modify(id, creature: Creature) -> Creature:
+def modify(name: str, creature: Creature) -> Creature:
     qry = """update creature
              set country=:country,
-                 name=:name,
-                 description=:description,
-                 area=:area,
-                 aka=:aka
+             name=:name,
+             description=:description,
+             area=:area,
+             aka=:aka
              where name=:name_orig"""
     params = model_to_dict(creature)
     params["name_orig"] = creature.name
-    _ = curs.execute(qry, params)
+    curs.execute(qry, params)
     conn.commit()
-    return get_one(creature.name)
+    if curs.rowcount == 1:
+        return get_one(creature.name)
+    else:
+        raise Missing(msg=f"Creature {name} not found")
 
-def delete(creature: Creature) -> None:
+def delete(name: str):
+    if not name: return False
     qry = "delete from creature where name = :name"
-    params = {"name": creature.name}
-    res = curs.execute(qry, params)
+    params = {"name": name}
+    curs.execute(qry, params)
     conn.commit()
+    if curs.rowcount != 1:
+        raise Missing(msg=f"Creature {name} not found")
